web-animations/animation-group/animation-group-children.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>AnimationGroup children attribute test</title>
<meta name="assert" content="The list of child animation nodes in the group">
<link rel="help" href="http://w3c.github.io/web-animations/#dom-animationgroup-children">
<link rel="help" href="http://w3c.github.io/web-animations/#child-animation-node">
<link rel="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<link rel="author" title="Aleksei Yu. Semenov" href="mailto:a.semenov@unipro.ru">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../testcommon.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
<body>
<div id="log"></div>
<script>
test(function() {
    var ctors = [AnimationGroup, AnimationSequence];
    ctors.forEach(function(ctor) {
        var group = new ctor([]);

        assert_array_equals(group.children, [], ctor.name + '.children should be empty');
    });
}, 'Test AnimationGroup.children returns empty list for animation group without children');

test(function() {
    var test = this;
    var ctors = [AnimationGroup, AnimationSequence];
    ctors.forEach(function(ctor) {
        var node = new Animation(createDiv(test), KEYFRAMES, ANIMATION_DURATION);
        var children = [node];
        var group = new ctor(children);

        assert_array_equals(group.children, children, ctor.name + '.children should correspond ' +
            'the first constructor argument');
    });
}, 'Test AnimationGroup.children returns the list of child animation nodes in the group. ' +
    'Test group with single child');

test(function() {
    var test = this;
    var ctors = [AnimationGroup, AnimationSequence];
    ctors.forEach(function(ctor) {
        var node1 = new Animation(createDiv(test), KEYFRAMES, ANIMATION_DURATION);
        var node2 = new AnimationGroup([]);
        var node3 = new Animation(createDiv(test), KEYFRAMES, ANIMATION_DURATION);
        var node4 = new AnimationSequence([]);
        var children = [node1, node2, node3, node4];
        var group = new ctor(children);

        assert_array_equals(group.children, children, ctor.name + '.children should correspond ' +
            'the first constructor argument');
    });
}, 'Test AnimationGroup.children returns the list of child animation nodes in the group. ' +
    'Test group with different children: Animations and AnimationGroups');

test(function() {
    var test = this;
    var ctors = [AnimationGroup, AnimationSequence];
    ctors.forEach(function(ctor) {
        var node1 = new Animation(createDiv(test), KEYFRAMES, ANIMATION_DURATION);
        var node2 = new Animation(createDiv(test), KEYFRAMES, ANIMATION_DURATION);
        var node3 = new Animation(createDiv(test), KEYFRAMES, ANIMATION_DURATION);
        var node4 = new ctor([node1, node2]);
        var node5 = new Animation(createDiv(test), KEYFRAMES, ANIMATION_DURATION);
        var children = [node3, node4, node5];
        var group = new ctor(children);

        assert_array_equals(group.children, children, ctor.name + '.children should correspond ' +
            'the first constructor argument');
        assert_array_equals(node4.children, [node1, node2], ctor.name + '.children should ' +
            'return the first constructor argument');
    });
}, 'Test AnimationGroup.children returns the list of child animation nodes in the group. ' +
    'Test low-level node in the tree');
</script>
</body>
